# public/openapi.generated.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Life@USTC API",
    "version": "1.0.0",
    "description": "OpenAPI document generated from Next.js routes by next-openapi-gen"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/api/uploads": {
      "get": {
        "operationId": "get-api-uploads",
        "summary": "List uploads of current user.",
        "description": "",
        "tags": ["Api"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/uploadsListResponseSchema"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post-api-uploads",
        "summary": "Create a signed upload URL.",
        "description": "",
        "tags": ["Api"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/uploadCreateRequestSchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/uploadCreateResponseSchema"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "uploadsListResponseSchema": {
        "type": "object",
        "properties": {
          "maxFileSizeBytes": {
            "type": "integer"
          },
          "quotaBytes": {
            "type": "integer"
          },
          "uploads": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/uploadSummarySchema"
            }
          },
          "usedBytes": {
            "type": "integer"
          }
        },
        "required": ["maxFileSizeBytes", "quotaBytes", "uploads", "usedBytes"]
      },
      "uploadCreateRequestSchema": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string",
            "minLength": 1
          },
          "contentType": {
            "type": "string"
          },
          "size": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ]
          }
        },
        "required": ["filename", "size"]
      },
      "uploadCreateResponseSchema": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "maxFileSizeBytes": {
            "type": "integer"
          },
          "quotaBytes": {
            "type": "integer"
          },
          "usedBytes": {
            "type": "integer"
          }
        },
        "required": [
          "key",
          "url",
          "maxFileSizeBytes",
          "quotaBytes",
          "usedBytes"
        ]
      },
...
      "dateTimeSchema": {
        "type": "string"
      },
...
      "uploadSummarySchema": {
        "type": "object",
        "properties": {
          "createdAt": {
            "$ref": "#/components/schemas/dateTimeSchema"
          }
        },
        "required": ["createdAt"]
      },
...
    },
...
  }
}